Moved ProffesionsController to DoctorController
